Conflicts:
	cave/com.raytheon.viz.hydro/src/com/raytheon/viz/hydro/timeseries/TimeSeriesDlg.java
	cave/com.raytheon.viz.radar/src/com/raytheon/viz/radar/rsc/graphic/RadarGraphicsPage.java


Former-commit-id: 86ef513f00aaa0c06f66088cf2e67f25da788e57

+++ edexOsgi/com.raytheon.uf.edex.damagepath/src/com/raytheon/uf/edex/damagepath/handler/ExportToLdadHandler.java
@@ -25,8 +25,11 @@ import java.nio.file.Path;
 import java.nio.file.attribute.FileAttribute;
 import java.nio.file.attribute.PosixFilePermission;
 import java.nio.file.attribute.PosixFilePermissions;
+import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.Set;
+
+import org.apache.commons.lang3.StringUtils;
 
 import com.raytheon.uf.common.damagepath.request.ExportToLdadRequest;
 import com.raytheon.uf.common.serialization.comm.IRequestHandler;
@@ -47,6 +50,8 @@ import com.raytheon.uf.common.util.RunProcess;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Jun 08, 2015  #4355     dgilling    Initial creation
+ * Mar 10, 2016  #5288     dgilling    Add ability to run command on LDAD host
+ *                                     via SSH after file has been copied.
  * 
  * </pre>
  * 
@@ -58,11 +63,11 @@ public final class ExportToLdadHandler implements
         IRequestHandler<ExportToLdadRequest> {
 
     private final IUFStatusHandler statusHandler = UFStatus
-            .getHandler(ExportToLdadHandler.class);
+            .getHandler(getClass());
 
     private static final String TEMP_FILE_NAME = "damage-path-tool";
 
-    private static final String FILE_EXTENSION = ".js";
+    private static final String FILE_EXTENSION = ".json";
 
     private static final FileAttribute<Set<PosixFilePermission>> TEMP_FILE_PERMISSIONS = PosixFilePermissions
             .asFileAttribute(EnumSet.of(PosixFilePermission.OWNER_READ,
@@ -74,44 +79,67 @@ public final class ExportToLdadHandler implements
     private static final String EXPORT_FILE_NAME_FMT = "%s.DPT.%2tY%<tm%<td.%<tH%<tM%<tS"
             + FILE_EXTENSION;
 
-    private static final String ERROR_MSG_FORMAT = "Command [%s] returned non-zero exit code %d. Process output: %s";
+    private static final String ERROR_MSG_FORMAT = "Command %s returned non-zero exit code %d. Process output: %s";
 
-    private final String scpHost;
+    private final String host;
 
-    private final String scpUser;
+    private final String user;
 
     private final String scpDestination;
 
+    private final String postProcessCommand;
+
     public ExportToLdadHandler(final ExportToLdadJobConfig config) {
-        this.scpHost = config.getHost();
-        this.scpUser = config.getUser();
+        this.host = config.getHost();
+        this.user = config.getUser();
         this.scpDestination = config.getDestinationPath();
+        this.postProcessCommand = config.getPostProcessCommand();
     }
 
     @Override
     public String handleRequest(ExportToLdadRequest request) throws Exception {
-        String retVal = null;
-
-        Path srcFile = writeTempFile(request.getData());
-        String command = buildCommand(srcFile.toString(), request.getSiteId());
-
-        RunProcess p = RunProcess.getRunProcess();
-        p.exec(command);
-        int returnCode = p.waitFor();
-        if (returnCode != 0) {
-            retVal = String.format(ERROR_MSG_FORMAT, command, returnCode,
-                    p.getStderr());
-            statusHandler.error(retVal);
-        }
+        Path srcFile = null;
 
         try {
-            Files.delete(srcFile);
-        } catch (Exception e) {
-            statusHandler.handle(Priority.WARN,
-                    "Could not delete temporary file.", e);
+            srcFile = writeTempFile(request.getData());
+            String destFileName = getDestFileName(request.getSiteId());
+
+            String[] command = buildScpCommand(srcFile.toString(), destFileName);
+            String errorMsg = runSystemCommand(command);
+            if (!errorMsg.isEmpty()) {
+                statusHandler.error(errorMsg);
+                return errorMsg;
+            }
+
+            /*
+             * Executing a remote command after sending the file to LDAD is
+             * optional.
+             */
+            if (StringUtils.isNotBlank(postProcessCommand)) {
+                command = buildSshCommand(destFileName);
+                errorMsg = runSystemCommand(command);
+                if (!errorMsg.isEmpty()) {
+                    statusHandler.error(errorMsg);
+                    return errorMsg;
+                }
+            }
+        } finally {
+            if (srcFile != null) {
+                /*
+                 * We want to catch all exceptions from deleting the temp file
+                 * because it does not warrant informing the caller that it
+                 * failed. we can simply log a warning.
+                 */
+                try {
+                    Files.delete(srcFile);
+                } catch (Exception e) {
+                    statusHandler.handle(Priority.WARN,
+                            "Could not delete temporary file.", e);
+                }
+            }
         }
 
-        return retVal;
+        return StringUtils.EMPTY;
     }
 
     private Path writeTempFile(final byte[] contents) throws IOException {
@@ -120,16 +148,43 @@ public final class ExportToLdadHandler implements
         return Files.write(tempFile, contents);
     }
 
-    private String buildCommand(final String sourceFile, final String siteID) {
-        StringBuilder ldadCommand = new StringBuilder("scp -q");
-        ldadCommand.append(' ');
-        ldadCommand.append(sourceFile);
-        ldadCommand.append(' ');
-        ldadCommand.append(scpUser).append('@').append(scpHost);
-        ldadCommand.append(':');
-        ldadCommand.append(scpDestination);
-        ldadCommand.append(String.format(EXPORT_FILE_NAME_FMT, siteID,
-                TimeUtil.newDate()));
-        return ldadCommand.toString();
+    private String getDestFileName(final String siteID) {
+        return String.format(EXPORT_FILE_NAME_FMT, siteID, TimeUtil.newDate());
+    }
+
+    private String[] buildScpCommand(final String sourceFile,
+            final String destFile) {
+        StringBuilder remoteDestination = new StringBuilder();
+        remoteDestination.append(user).append('@').append(host);
+        remoteDestination.append(':');
+        remoteDestination.append(scpDestination);
+        remoteDestination.append(destFile);
+
+        return new String[] { "scp", "-q", sourceFile,
+                remoteDestination.toString() };
+    }
+
+    private String[] buildSshCommand(final String destFile) {
+        StringBuilder connectionString = new StringBuilder();
+        connectionString.append(user).append('@').append(host);
+
+        StringBuilder remoteCommand = new StringBuilder();
+        remoteCommand.append(postProcessCommand).append(' ')
+                .append(scpDestination).append(destFile);
+
+        return new String[] { "ssh", "-q", connectionString.toString(),
+                remoteCommand.toString() };
+    }
+
+    private String runSystemCommand(String[] commandAndArgs) throws IOException {
+        RunProcess p = RunProcess.getRunProcess();
+        p.exec(commandAndArgs);
+        int returnCode = p.waitFor();
+        if (returnCode != 0) {
+            return String.format(ERROR_MSG_FORMAT,
+                    Arrays.toString(commandAndArgs), returnCode, p.getStderr());
+        }
+
+        return StringUtils.EMPTY;
     }
 }
